+++ crud/insert.lua
@@ -7,7 +7,6 @@ local utils = require('crud.common.utils')
 local sharding = require('crud.common.sharding')
 local dev_checks = require('crud.common.dev_checks')
 local schema = require('crud.common.schema')
-local bucket_ref_unref = require('crud.common.sharding.bucket_ref_unref')
 
 local InsertError = errors.new_class('InsertError', {capture_stack = false})
 
@@ -43,29 +42,15 @@ local function insert_on_storage(space_name, tuple, opts)
         return nil, err
     end
 
-    local bucket_id = tuple[utils.get_bucket_id_fieldno(space)]
-    local ref_ok, bucket_ref_err, unref = bucket_ref_unref.bucket_refrw(bucket_id, space.engine)
-
-    if not ref_ok then
-        return nil, bucket_ref_err
-    end
-
     -- add_space_schema_hash is true only in case of insert_object
     -- the only one case when reloading schema can avoid insert error
     -- is flattening object on router
-    local result = schema.wrap_func_result(space, space.insert, {
+    return schema.wrap_func_result(space, space.insert, {
         add_space_schema_hash = opts.add_space_schema_hash,
         field_names = opts.fields,
         noreturn = opts.noreturn,
         fetch_latest_metadata = opts.fetch_latest_metadata,
     }, space, tuple)
-
-    local unref_ok, err_unref = unref(bucket_id, space.engine)
-    if not unref_ok then
-        return nil, err_unref
-    end
-
-    return result
 end
 
 insert.storage_api = {[INSERT_FUNC_NAME] = insert_on_storage}
@@ -115,7 +100,7 @@ local function call_insert_on_router(vshard_router, space_name, original_tuple, 
         timeout = opts.timeout,
     }
 
-    local storage_result, err = call.single(vshard_router,
+    local storage_result, err = call.single_direct(vshard_router,
         sharding_data.bucket_id, CRUD_INSERT_FUNC_NAME,
         {space_name, tuple, insert_on_storage_opts},
         call_opts