+++ crud/insert_many.lua
@@ -58,7 +58,7 @@ local function insert_many_on_storage(space_name, tuples, opts)
     local errs = {}
     local replica_schema_version = nil
 
-    local ref_ok, bucket_ref_err, unref = bucket_ref_unref.bucket_refrw_many(bucket_ids, space.engine)
+    local ref_ok, bucket_ref_err, unref = bucket_ref_unref.bucket_refrw_many(bucket_ids)
     if not ref_ok then
         table.insert(errs, bucket_ref_err)
         return nil, errs, replica_schema_version
@@ -102,7 +102,7 @@ local function insert_many_on_storage(space_name, tuples, opts)
 
                 if opts.rollback_on_error == true then
                     box.rollback()
-                    local unref_ok, bucket_unref_err = unref(bucket_ids, space.engine)
+                    local unref_ok, bucket_unref_err = unref(bucket_ids)
                     if not unref_ok then
                         table.insert(errs, bucket_unref_err)
                         return nil, errs, replica_schema_version
@@ -116,7 +116,7 @@ local function insert_many_on_storage(space_name, tuples, opts)
                 end
 
                 box.commit()
-                local unref_ok, bucket_unref_err = unref(bucket_ids, space.engine)
+                local unref_ok, bucket_unref_err = unref(bucket_ids)
                 if not unref_ok then
                     table.insert(errs, bucket_unref_err)
                     return nil, errs, replica_schema_version
@@ -132,7 +132,7 @@ local function insert_many_on_storage(space_name, tuples, opts)
     if next(errs) ~= nil then
         if opts.rollback_on_error == true then
             box.rollback()
-            local unref_ok, bucket_unref_err = unref(bucket_ids, space.engine)
+            local unref_ok, bucket_unref_err = unref(bucket_ids)
             if not unref_ok then
                 table.insert(errs, bucket_unref_err)
                 return nil, errs, replica_schema_version
@@ -146,7 +146,7 @@ local function insert_many_on_storage(space_name, tuples, opts)
         end
 
         box.commit()
-        local unref_ok, bucket_unref_err = unref(bucket_ids, space.engine)
+        local unref_ok, bucket_unref_err = unref(bucket_ids)
         if not unref_ok then
             table.insert(errs, bucket_unref_err)
             return nil, errs, replica_schema_version
@@ -156,7 +156,7 @@ local function insert_many_on_storage(space_name, tuples, opts)
     end
 
     box.commit()
-    local unref_ok, bucket_unref_err = unref(bucket_ids, space.engine)
+    local unref_ok, bucket_unref_err = unref(bucket_ids)
     if not unref_ok then
         table.insert(errs, bucket_unref_err)
         return nil, errs, replica_schema_version