+++ crud/replace_many.lua
@@ -65,7 +65,7 @@ local function replace_many_on_storage(space_name, tuples, opts)
         bucket_ids[tuple[utils.get_bucket_id_fieldno(space)]] = true
     end
 
-    local ref_ok, bucket_ref_err, unref = bucket_ref_unref.bucket_refrw_many(bucket_ids, space.engine)
+    local ref_ok, bucket_ref_err, unref = bucket_ref_unref.bucket_refrw_many(bucket_ids)
     if not ref_ok then
         table.insert(errs, bucket_ref_err)
         return nil, errs, replica_schema_version
@@ -105,7 +105,7 @@ local function replace_many_on_storage(space_name, tuples, opts)
 
                 if opts.rollback_on_error == true then
                     box.rollback()
-                    local unref_ok, bucket_unref_err = unref(bucket_ids, space.engine)
+                    local unref_ok, bucket_unref_err = unref(bucket_ids)
                     if not unref_ok then
                         table.insert(errs, bucket_unref_err)
                         return nil, errs, replica_schema_version
@@ -118,7 +118,7 @@ local function replace_many_on_storage(space_name, tuples, opts)
                 end
 
                 box.commit()
-                local unref_ok, bucket_unref_err = unref(bucket_ids, space.engine)
+                local unref_ok, bucket_unref_err = unref(bucket_ids)
                 if not unref_ok then
                     table.insert(errs, bucket_unref_err)
                     return nil, errs, replica_schema_version
@@ -133,7 +133,7 @@ local function replace_many_on_storage(space_name, tuples, opts)
     if next(errs) ~= nil then
         if opts.rollback_on_error == true then
             box.rollback()
-            local unref_ok, bucket_unref_err = unref(bucket_ids, space.engine)
+            local unref_ok, bucket_unref_err = unref(bucket_ids)
             if not unref_ok then
                 table.insert(errs, bucket_unref_err)
                 return nil, errs, replica_schema_version
@@ -147,7 +147,7 @@ local function replace_many_on_storage(space_name, tuples, opts)
         end
 
         box.commit()
-        local unref_ok, bucket_unref_err = unref(bucket_ids, space.engine)
+        local unref_ok, bucket_unref_err = unref(bucket_ids)
         if not unref_ok then
             table.insert(errs, bucket_unref_err)
             return nil, errs, replica_schema_version
@@ -157,7 +157,7 @@ local function replace_many_on_storage(space_name, tuples, opts)
     end
 
     box.commit()
-    local unref_ok, bucket_unref_err = unref(bucket_ids, space.engine)
+    local unref_ok, bucket_unref_err = unref(bucket_ids)
     if not unref_ok then
         table.insert(errs, bucket_unref_err)
         return nil, errs, replica_schema_version