+++ test/unit/stats_test.lua
@@ -45,7 +45,6 @@ local function enable_stats(g, params)
         params = table.deepcopy(params)
         params.backend = nil
         params.backend_cfg = nil
-        params.safe_mode = nil
     end
     g.router:eval("stats_module.enable(...)", { params })
 end